Add shared email availability hook
